Move percentages to their own actions
