+++ lib/flipper/ui/actions/percentage_of_actors_gate.rb
@@ -1,0 +1,27 @@
+require 'flipper/ui/action'
+require 'flipper/ui/decorators/feature'
+
+module Flipper
+  module UI
+    module Actions
+      class PercentageOfActorsGate < UI::Action
+        route %r{features/[^/]*/percentage_of_actors/?\Z}
+
+        def post
+          feature_name = Rack::Utils.unescape(request.path.split("/")[-2])
+          feature = flipper[feature_name.to_sym]
+          @feature = Decorators::Feature.new(feature)
+
+          begin
+            feature.enable_percentage_of_actors params["value"]
+          rescue ArgumentError => exception
+            error = Rack::Utils.escape("Invalid percentage of actors value: #{exception.message}")
+            redirect_to("/features/#{@feature.key}?error=#{error}")
+          end
+
+          redirect_to "/features/#{@feature.key}"
+        end
+      end
+    end
+  end
+end
