+++ lib/flipper/ui/middleware.rb
@@ -50,6 +50,8 @@ module Flipper
         @action_collection.add UI::Actions::AddActor
         @action_collection.add UI::Actions::AddGroup
         @action_collection.add UI::Actions::BooleanGate
+        @action_collection.add UI::Actions::PercentageOfTimeGate
+        @action_collection.add UI::Actions::PercentageOfActorsGate
         @action_collection.add UI::Actions::Gate
 
         # Catch all, always last.
