+++ spec/flipper/ui/actions/boolean_gate_spec.rb
@@ -4,7 +4,7 @@ require 'flipper'
 require 'flipper/adapters/memory'
 require 'flipper/ui/actor'
 
-describe Flipper::UI::Actions::Gate do
+describe Flipper::UI::Actions::BooleanGate do
   include Rack::Test::Methods
 
   let(:adapter) { Flipper::Adapters::Memory.new }
